Fix test-utils structure
